Adds custom registration step
