+++ CardinalKit-Example/CardinalKit/Components/Onboarding/OnboardingViewController.swift
@@ -45,23 +45,8 @@ struct OnboardingViewController: UIViewControllerRepresentable {
         /* ***********************************************************
         * STEP (3): Get the user's personal information for registration
          ************************************************************* */
-        
-        let nameFormStep = ORKFormStep(identifier: "nameFormStep", title: "Registration", text: "Please provide your information for creating your user account.")
-        
-        let firstNameAnswerFormat = ORKTextAnswerFormat(maximumLength: 20)
-        firstNameAnswerFormat.multipleLines = false
-        
-        let lastNameAnswerFormat = ORKTextAnswerFormat(maximumLength: 20)
-        lastNameAnswerFormat.multipleLines = false
-        
-        let nameSectionTitle = ORKFormItem(sectionTitle: nil)
-        let firstNameFormItem = ORKFormItem(identifier: "firstNameFormItem", text: "First Name", answerFormat: firstNameAnswerFormat)
-        let lastNameFormItem = ORKFormItem(identifier: "lastNameFormItem", text: "Last Name", answerFormat: lastNameAnswerFormat)
-        let dobAnswerFormat = ORKAnswerFormat.dateAnswerFormat(withDefaultDate: nil, minimumDate: nil, maximumDate: Date(), calendar: nil)
-        let dobFormItem = ORKFormItem(identifier: "dobFormItem", text: "Date of Birth", answerFormat: dobAnswerFormat)
-        
-        nameFormStep.formItems = [nameSectionTitle, firstNameFormItem, lastNameFormItem, dobFormItem]
-        
+        // see CustomRegistrationStep to configure
+        let registrationStep = CustomRegistrationStep.registrationForm
         
         /* **************************************************************
         *  STEP (4): get permission to collect HealthKit data
@@ -129,7 +114,7 @@ struct OnboardingViewController: UIViewControllerRepresentable {
         **************************************************************/
         
         // given intro steps that the user should review and consent to
-        let introSteps: [ORKStep] = [consentStep, reviewConsentStep, nameFormStep]
+        let introSteps: [ORKStep] = [consentStep, reviewConsentStep, registrationStep]
         
         // and steps regarding login / security
         let emailVerificationSteps = loginSteps + [passcodeStep, healthDataStep, healthRecordsStep, completionStep]
